Progress bars are now themed but still very basic. Considering removing progress bars and multithreading support since it causes many problems or at least maintaining it on a different branch.
