+++ src/cryogen/Cryptogen.java
@@ -23,9 +23,6 @@ import java.awt.datatransfer.StringSelection;
 import java.awt.datatransfer.Transferable;
 import java.io.*;
 import java.net.URL;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.*;
 import java.util.List;
 
@@ -291,7 +288,7 @@ public class Cryptogen implements Initializable
 
             encryptFiles(files);
 
-            Alert encryptionInformation = new Alert(Alert.AlertType.INFORMATION, message + method + message2);
+            /*Alert encryptionInformation = new Alert(Alert.AlertType.INFORMATION, message + method + message2);
             encryptionInformation.setGraphic(new ImageView(this.getClass().getResource("/icons/success32.png").toString()));
             Button okButton = (Button) encryptionInformation.getDialogPane().lookupButton(ButtonType.OK);
             okButton.setText("OK");
@@ -305,7 +302,7 @@ public class Cryptogen implements Initializable
             if (!ButtonType.OK.equals(closeResponse.get()))
             {
                 event.consume();
-            }
+            }*/
         }
         catch (NoFilesAttachedException ex)
         {
@@ -419,7 +416,7 @@ public class Cryptogen implements Initializable
 
             decryptFiles(files);
 
-            Alert decryptionInformation = new Alert(Alert.AlertType.INFORMATION, message + method);
+            /*Alert decryptionInformation = new Alert(Alert.AlertType.INFORMATION, message + method);
             decryptionInformation.setGraphic(new ImageView(this.getClass().getResource("/icons/success32.png").toString()));
             decryptionInformation.initModality(Modality.APPLICATION_MODAL);
             decryptionInformation.initOwner(getCurrentStage());
@@ -435,7 +432,7 @@ public class Cryptogen implements Initializable
             if (!ButtonType.OK.equals(closeResponse.get()))
             {
                 event.consume();
-            }
+            }*/
         }
         catch (NoFilesAttachedException ex)
         {
